pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evalkit"
version = "0.1.0"
description = "CLI eval harness for LLMs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "respx>=0.20",
]

[project.scripts]
evalkit = "evalkit.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/evalkit"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
